Thu, Oct 20, 2022, 10:20 PM +00:00

+++ scores.c
@@ -12,5 +12,5 @@ int main(void)
         scores[i] = get_int("Score ");
     }
 
-    printf("Average %d\n", (scores[0] + scores[1] + scores[2]) / n);
+    printf("Average %f\n", (scores[0] + scores[1] + scores[2]) / (float) n);
 }
